Fix float conversion

+++ rider-state-validation/app.py
@@ -7,8 +7,8 @@ def handler(event, context):
     print(json.dumps(event))
     
     rider_id = event['rider']['rider_id']
-    lat = int(event['rider']['lat'])
-    long = int(event['rider']['long'])
+    lat = float(event['rider']['lat'])
+    long = float(event['rider']['long'])
 
     if lat >= 0 and long >= 0:
         return {
